requirements.txt
pandas>=2.0.0
numpy
scipy
yfinance>=0.2.31
pandas-ta-classic
python-telegram-bot>=20.0
python-dotenv
pytz
requests
pytest
pytest-asyncio
scikit-learn
joblib
google-genai
gspread==6.0.0
oauth2client==4.1.3
fastapi
uvicorn
